distributed currency stuff between existing modules
